#910 Enable rule S104 and split graph
