+++ src/Configuration/test/ConfigServer.Test/.editorconfig
@@ -1,0 +1,2 @@
+[*.{cs,vb}]
+dotnet_diagnostic.S104.severity = none
