+++ src/Management/src/Endpoint/HeapDump/DotNetHeapDump/Graph.cs
@@ -25,63 +25,63 @@ namespace Graphs
 {
     /// <summary>
     /// A graph is representation of a node-arc graph.    It tries to be very space efficient.   It is a little
-    /// more complex than the  most basic node-arc graph in that each node can have a code:NodeType associated with it 
-    /// that contains information that is shared among many nodes.   
-    /// 
-    /// While the 'obvious' way of representing a graph is to have a 'Node' object that has arcs, we don't do this. 
+    /// more complex than the  most basic node-arc graph in that each node can have a code:NodeType associated with it
+    /// that contains information that is shared among many nodes.
+    ///
+    /// While the 'obvious' way of representing a graph is to have a 'Node' object that has arcs, we don't do this.
     /// Instead each node is give an unique code:NodeIndex which represents the node and each node has a list of
     /// NodeIndexes for each of the children.   Using indexes instead of object pointers is valuable because
-    /// 
+    ///
     ///     * You can save 8 bytes (on 32 bit) of .NET object overhead and corresponding cost at GC time by using
     ///       indexes.   This is significant because there can be 10Meg of objects, so any expense adds up
     ///     * Making the nodes be identified by index is more serialization friendly.   It is easier to serialize
-    ///       the graph if it has this representation.  
+    ///       the graph if it has this representation.
     ///     * It easily allows 3rd parties to 'attach' their own information to each node.  All they need is to
     ///       create an array of the extra information indexed by NodeIndex.   The 'NodeIndexLimit' is designed
-    ///       specifically for this purpose.  
-    ///       
-    /// Because we anticipate VERY large graphs (e.g. dumps of the GC heap) the representation for the nodes is 
+    ///       specifically for this purpose.
+    ///
+    /// Because we anticipate VERY large graphs (e.g. dumps of the GC heap) the representation for the nodes is
     /// very space efficient and we don't have code:Node class object for most of the nodes in the graph.  However
-    /// it IS useful to have code:Node objects for the nodes that are being manipulated locally.  
+    /// it IS useful to have code:Node objects for the nodes that are being manipulated locally.
     ///
     /// To avoid creating lots of code:Node objects that die quickly the API adopts the convention that the
     /// CALLer provides a code:Node class as 'storage' when the API needs to return a code:Node.   That way
     /// APIs that return code:Node never allocate.    This allows most graph algorithms to work without having
     /// to allocate more than a handful of code:Node classes, reducing overhead.   You allocate these storage
     /// nodes with the code:Graph.AllocNodeStorage call
-    /// 
+    ///
     /// Thus the basic flow is you call code:Graph.AllocNodeStorage to allocate storage, then call code:Graph.GetRoot
     /// to get your first node.  If you need to provide additional information about the nodes, you can allocate an auxiliary
-    /// array of Size code:Graph.NodeIndexLimit to hold it (for example a 'visited' bit).   Then repeatedly call 
+    /// array of Size code:Graph.NodeIndexLimit to hold it (for example a 'visited' bit).   Then repeatedly call
     /// code:Node.GetFirstChild, code:Node.GetNextChild to get the children of a node to traverse the graph.
-    /// 
+    ///
     /// OVERHEAD
     ///
     ///     1) 4 bytes per Node for the pointer to the stream for the rest of the data (thus we can have at most 4Gig nodes)
-    ///     2) For each node, the number of children, the nodeId, and children are stored as compressed (relative) indexes 
+    ///     2) For each node, the number of children, the nodeId, and children are stored as compressed (relative) indexes
     ///        (figure 1 byte for # of children, 2 bytes per type id, 2-3 bytes per child)
     ///     3) Variable length nodes also need a compressed int for the Size of the node (1-3 bytes)
-    ///     4) Types store the name (2 bytes per character), and Size (4 bytes), but typically don't dominate 
-    ///        Size of graph.  
+    ///     4) Types store the name (2 bytes per character), and Size (4 bytes), but typically don't dominate
+    ///        Size of graph.
     ///
     /// Thus roughly 7 bytes per node + 3 bytes per reference.   Typically nodes on average have 2-3 references, so
-    /// figure 13-16 bytes per node.  That gives you 125 Million nodes in a 2 Gig of Memory. 
-    /// 
+    /// figure 13-16 bytes per node.  That gives you 125 Million nodes in a 2 Gig of Memory.
+    ///
     /// The important point here is that representation of a node is always smaller than the Memory it represents, and
-    /// and often significantly smaller (since it does not hold non-GC data, null pointers and even non-null pointers 
-    /// are typically half the Size).   For 64 bit heaps, the Size reduction is even more dramatic.  
-    /// 
+    /// and often significantly smaller (since it does not hold non-GC data, null pointers and even non-null pointers
+    /// are typically half the Size).   For 64 bit heaps, the Size reduction is even more dramatic.
+    ///
     /// see code:Graph.SizeOfGraphDescription to determine the overhead for any particular graph.
-    /// 
+    ///
     /// </summary>
     [ExcludeFromCodeCoverage]
     internal abstract class Graph : IFastSerializable, IFastSerializableVersion
     {
         /// <summary>
-        /// Given an arbitrary code:NodeIndex that identifies the node, Get a code:Node object.  
-        /// 
-        /// This routine does not allocated but uses the space passed in by 'storage.  
-        /// 'storage' should be allocated with code:AllocNodeStorage, and should be aggressively reused.  
+        /// Given an arbitrary code:NodeIndex that identifies the node, Get a code:Node object.
+        ///
+        /// This routine does not allocated but uses the space passed in by 'storage.
+        /// 'storage' should be allocated with code:AllocNodeStorage, and should be aggressively reused.
         /// </summary>
         public Node GetNode(NodeIndex nodeIndex, Node storage)
         {
@@ -90,17 +90,17 @@ namespace Graphs
             return storage;
         }
         /// <summary>
-        /// returns true if SetNode has been called on this node (it is not an undefined object).  
-        /// TODO FIX NOW used this instead of the weird if node index grows technique. 
+        /// returns true if SetNode has been called on this node (it is not an undefined object).
+        /// TODO FIX NOW used this instead of the weird if node index grows technique.
         /// </summary>
         public bool IsDefined(NodeIndex nodeIndex) { return m_nodes[(int)nodeIndex] != m_undefinedObjDef; }
         /// <summary>
-        /// Given an arbitrary code:NodeTypeIndex that identifies the nodeId of the node, Get a code:NodeType object.  
-        /// 
-        /// This routine does not allocated but overwrites the space passed in by 'storage'.  
-        /// 'storage' should be allocated with code:AllocNodeTypeStorage, and should be aggressively reused.  
-        /// 
-        /// Note that this routine does not get used much, instead Node.GetType is normal way of getting the nodeId.  
+        /// Given an arbitrary code:NodeTypeIndex that identifies the nodeId of the node, Get a code:NodeType object.
+        ///
+        /// This routine does not allocated but overwrites the space passed in by 'storage'.
+        /// 'storage' should be allocated with code:AllocNodeTypeStorage, and should be aggressively reused.
+        ///
+        /// Note that this routine does not get used much, instead Node.GetType is normal way of getting the nodeId.
         /// </summary>
         public NodeType GetType(NodeTypeIndex nodeTypeIndex, NodeType storage)
         {
@@ -132,7 +132,7 @@ namespace Graphs
         /// </summary>
         public NodeIndex NodeIndexLimit { get { return (NodeIndex)m_nodes.Count; } }
         /// <summary>
-        /// Same as NodeIndexLimit, just cast to an integer.  
+        /// Same as NodeIndexLimit, just cast to an integer.
         /// </summary>
         public int NodeCount { get { return (int)m_nodes.Count; } }
         /// <summary>
@@ -141,11 +141,11 @@ namespace Graphs
         /// </summary>
         public NodeTypeIndex NodeTypeIndexLimit { get { return (NodeTypeIndex)m_types.Count; } }
         /// <summary>
-        /// Same as NodeTypeIndex cast as an integer.  
+        /// Same as NodeTypeIndex cast as an integer.
         /// </summary>
         public int NodeTypeCount { get { return m_types.Count; } }
         /// <summary>
-        /// When a Node is created, you specify how big it is.  This the sum of all those sizes.  
+        /// When a Node is created, you specify how big it is.  This the sum of all those sizes.
         /// </summary>
         public long TotalSize { get { return m_totalSize; } }
         /// <summary>
@@ -162,16 +162,16 @@ namespace Graphs
         /// </summary>
         protected const int SegmentSize = 8_192;
 
-        // Creation methods.  
+        // Creation methods.
         /// <summary>
         /// Create a new graph from 'nothing'.  Note you are not allowed to read from the graph
-        /// until you execute 'AllowReading'.  
-        /// 
+        /// until you execute 'AllowReading'.
+        ///
         /// You can actually continue to write after executing 'AllowReading' however you should
         /// any additional nodes you write should not be accessed until you execute 'AllowReading'
-        /// again.  
-        /// 
-        /// TODO I can eliminate the need for AllowReading.  
+        /// again.
+        ///
+        /// TODO I can eliminate the need for AllowReading.
         /// </summary>
         public Graph(int expectedNodeCount)
         {
@@ -187,9 +187,9 @@ namespace Graphs
         public NodeIndex RootIndex;
         /// <summary>
         /// Create a new nodeId with the given name and return its node nodeId index.   No interning is done (thus you can
-        /// have two distinct NodeTypeIndexes that have exactly the same name.  
-        /// 
-        /// By default the size = -1 which indicates we will set the type size to the first 'SetNode' for this type.  
+        /// have two distinct NodeTypeIndexes that have exactly the same name.
+        ///
+        /// By default the size = -1 which indicates we will set the type size to the first 'SetNode' for this type.
         /// </summary>
         public virtual NodeTypeIndex CreateType(string name, string moduleName = null, int size = -1)
         {
@@ -204,8 +204,8 @@ namespace Graphs
         /// <summary>
         /// Create a new node and return its index.   It is undefined until code:SetNode is called.   We allow undefined nodes
         /// because graphs have loops in them, and thus you need to refer to a node, before you know all the data in the node.
-        /// 
-        /// It is really expected that every node you did code:CreateNode on you also ultimately do a code:SetNode on.  
+        ///
+        /// It is really expected that every node you did code:CreateNode on you also ultimately do a code:SetNode on.
         /// </summary>
         /// <returns></returns>
         public virtual NodeIndex CreateNode()
@@ -231,8 +231,8 @@ namespace Graphs
         }
 
         /// <summary>
-        /// When a graph is constructed with the default constructor, it is in 'write Mode'  You can't read from it until 
-        /// you call 'AllowReading' which puts it in 'read mode'.  
+        /// When a graph is constructed with the default constructor, it is in 'write Mode'  You can't read from it until
+        /// you call 'AllowReading' which puts it in 'read mode'.
         /// </summary>
         public virtual void AllowReading()
         {
@@ -260,7 +260,7 @@ namespace Graphs
 #endif
         }
         /// <summary>
-        /// Used for debugging, returns the node Count and typeNode Count. 
+        /// Used for debugging, returns the node Count and typeNode Count.
         /// </summary>
         /// <returns></returns>
         public override string ToString()
@@ -268,11 +268,11 @@ namespace Graphs
             return string.Format("Graph of {0} nodes and {1} types.  Size={2:f3}MB SizeOfDescription={3:f3}MB",
                 NodeIndexLimit, NodeTypeIndexLimit, TotalSize / 1000000.0, SizeOfGraphDescription() / 1000000.0);
         }
-        // Performance 
+        // Performance
         /// <summary>
-        /// A pretty good estimate of the how many bytes of Memory it takes just to represent the graph itself. 
-        /// 
-        /// TODO: Currently this is only correct for the 32 bit version.  
+        /// A pretty good estimate of the how many bytes of Memory it takes just to represent the graph itself.
+        ///
+        /// TODO: Currently this is only correct for the 32 bit version.
         /// </summary>
         public virtual long SizeOfGraphDescription()
         {
@@ -300,35 +300,35 @@ namespace Graphs
 
         /* APIs for deferred lookup of type names */
         /// <summary>
-        /// Graph supports the ability to look up the names of a type at a later time.   You use this by 
-        /// calling this overload in which you give a type ID (e.g. an RVA) and a module index (from 
+        /// Graph supports the ability to look up the names of a type at a later time.   You use this by
+        /// calling this overload in which you give a type ID (e.g. an RVA) and a module index (from
         /// CreateModule) to this API.   If later you override the 'ResolveTypeName' delegate below
         /// then when type names are requested you will get back the typeID and module which you an
-        /// then use to look up the name (when you do have the PDB). 
-        /// 
-        /// The Module passed should be reused as much as possible to avoid bloated files.  
+        /// then use to look up the name (when you do have the PDB).
+        ///
+        /// The Module passed should be reused as much as possible to avoid bloated files.
         /// </summary>
         public NodeTypeIndex CreateType(int typeID, Module module, int size = -1, string typeNameSuffix = null)
         {
-            // make sure the m_types and m_deferedTypes arrays are in sync.  
+            // make sure the m_types and m_deferedTypes arrays are in sync.
             while (m_deferedTypes.Count < m_types.Count)
             {
                 m_deferedTypes.Add(new DeferedTypeInfo());
             }
 
             var ret = (NodeTypeIndex)m_types.Count;
-            // We still use the m_types array for the size. 
+            // We still use the m_types array for the size.
             m_types.Add(new TypeInfo { Size = size });
 
-            // but we put the real information into the m_deferedTypes.  
+            // but we put the real information into the m_deferedTypes.
             m_deferedTypes.Add(new DeferedTypeInfo { Module = module, TypeID = typeID, TypeNameSuffix = typeNameSuffix });
             Debug.Assert(m_deferedTypes.Count == m_types.Count);
             return ret;
         }
         /// <summary>
         /// In advanced scenarios you may not be able to provide a type name when you create the type.  YOu can pass null
-        /// for the type name to 'CreateType'   If you provide this callback, later you can provide the mapping from 
-        /// type index to name (e.g. when PDBs are available).    Note that this field is NOT serialized.   
+        /// for the type name to 'CreateType'   If you provide this callback, later you can provide the mapping from
+        /// type index to name (e.g. when PDBs are available).    Note that this field is NOT serialized.
         /// </summary>
         public Func<int, Module, string> ResolveTypeName { get; set; }
         /// <summary>
@@ -337,8 +337,8 @@ namespace Graphs
         public bool HasDeferedTypeNames { get { return m_deferedTypes.Count > 0; } }
 
         /* See GraphUtils class for more things you can do with a Graph. */
-        // TODO move these to GraphUtils. 
-        // Utility (could be implemented using public APIs).  
+        // TODO move these to GraphUtils.
+        // Utility (could be implemented using public APIs).
         public void BreadthFirstVisit(Action<Node> visitor)
         {
             var nodeStorage = AllocNodeStorage();
@@ -427,7 +427,7 @@ namespace Graphs
             m_nodes[(int)nodeIndex] = m_writer.GetLabel();
 
             Debug.Assert(sizeInBytes >= 0);
-            // We are going to assume that if this is negative it is because it is a large positive number.  
+            // We are going to assume that if this is negative it is because it is a large positive number.
             if (sizeInBytes < 0)
             {
                 sizeInBytes = int.MaxValue;
@@ -464,7 +464,7 @@ namespace Graphs
 
         /// <summary>
         /// ClearWorker does only that part of clear needed for this level of the hierarchy (and needs
-        /// to be done by the constructor too). 
+        /// to be done by the constructor too).
         /// </summary>
         private void ClearWorker()
         {
@@ -478,7 +478,7 @@ namespace Graphs
             m_nodes.Count = 0;
 
             // Create an undefined node, kind of gross because SetNode expects to have an entry
-            // in the m_nodes table, so we make a fake one and then remove it.  
+            // in the m_nodes table, so we make a fake one and then remove it.
             m_undefinedObjDef = m_writer.GetLabel();
             m_nodes.Add(m_undefinedObjDef);
             SetNode(0, CreateType("UNDEFINED"), 0, new GrowableArray<NodeIndex>());
@@ -488,25 +488,25 @@ namespace Graphs
 
         // To support very space efficient encodings, and to allow for easy serialization (persistence to file)
         // Types are given an index and their data is stored in a m_types array.  TypeInfo is the data in this
-        // array.  
+        // array.
         internal struct TypeInfo
         {
-            public string Name;                         // If DeferredTypeInfo.Module != null then this is a type name suffix.  
+            public string Name;                         // If DeferredTypeInfo.Module != null then this is a type name suffix.
             public int Size;
-            public string ModuleName;                   // The name of the module which contains the type (if known).  
+            public string ModuleName;                   // The name of the module which contains the type (if known).
         }
         internal struct DeferedTypeInfo
         {
             public int TypeID;
             public Module Module;                       // The name of the module which contains the type (if known).
-            public string TypeNameSuffix;               // if non-null it is added to the type name as a suffix.   
+            public string TypeNameSuffix;               // if non-null it is added to the type name as a suffix.
         }
 
         public virtual void ToStream(Serializer serializer)
         {
             serializer.Write(m_totalSize);
             serializer.Write((int)RootIndex);
-            // Write out the Types 
+            // Write out the Types
             serializer.Write(m_types.Count);
             for (int i = 0; i < m_types.Count; i++)
             {
@@ -515,15 +515,15 @@ namespace Graphs
                 serializer.Write(m_types[i].ModuleName);
             }
 
-            // Write out the Nodes 
+            // Write out the Nodes
             serializer.Write(m_nodes.Count);
             for (int i = 0; i < m_nodes.Count; i++)
             {
                 serializer.Write((int)m_nodes[i]);
             }
 
-            // Write out the Blob stream.  
-            // TODO this is inefficient.  Also think about very large files.  
+            // Write out the Blob stream.
+            // TODO this is inefficient.  Also think about very large files.
             int readerLen = (int)m_reader.Length;
             serializer.Write(readerLen);
             m_reader.Goto((StreamLabel)0);
@@ -533,17 +533,17 @@ namespace Graphs
             }
 
             // Are we writing a format for 1 or greater?   If so we can use the new (breaking) format, otherwise
-            // to allow old readers to read things, we give up on the new data.  
+            // to allow old readers to read things, we give up on the new data.
             if (1 <= ((IFastSerializableVersion)this).MinimumReaderVersion)
             {
                 // Because Graph has superclass, you can't add objects to the end of it (since it is not 'the end' of the object)
                 // which is a problem if we want to add new fields.  We could have had a worker object but another way of doing
                 // it is create a deferred (lazy region).   The key is that ALL readers know how to skip this region, which allows
-                // you to add new fields 'at the end' of the region (just like for sealed objects).  
+                // you to add new fields 'at the end' of the region (just like for sealed objects).
                 DeferedRegion expansion = new DeferedRegion();
                 expansion.Write(serializer, delegate
                 {
-                    // I don't need to use Tagged types for my 'first' version of this new region 
+                    // I don't need to use Tagged types for my 'first' version of this new region
                     serializer.Write(m_deferedTypes.Count);
                     for (int i = 0; i < m_deferedTypes.Count; i++)
                     {
@@ -553,8 +553,8 @@ namespace Graphs
                     }
 
                     // You can place tagged values in here always adding right before the WriteTaggedEnd
-                    // for any new fields added after version 1 
-                    serializer.WriteTaggedEnd(); // This insures tagged things don't read junk after the region.  
+                    // for any new fields added after version 1
+                    serializer.WriteTaggedEnd(); // This insures tagged things don't read junk after the region.
                 });
             }
         }
@@ -564,7 +564,7 @@ namespace Graphs
             deserializer.Read(out m_totalSize);
             RootIndex = (NodeIndex)deserializer.ReadInt();
 
-            // Read in the Types 
+            // Read in the Types
             TypeInfo info = new TypeInfo();
             int typeCount = deserializer.ReadInt();
             m_types = new GrowableArray<TypeInfo>(typeCount);
@@ -576,7 +576,7 @@ namespace Graphs
                 m_types.Add(info);
             }
 
-            // Read in the Nodes 
+            // Read in the Nodes
             int nodeCount = deserializer.ReadInt();
             m_nodes = new SegmentedList<StreamLabel>(SegmentSize, nodeCount);
 
@@ -585,8 +585,8 @@ namespace Graphs
                 m_nodes.Add((StreamLabel)(uint)deserializer.ReadInt());
             }
 
-            // Read in the Blob stream.  
-            // TODO be lazy about reading in the blobs.  
+            // Read in the Blob stream.
+            // TODO be lazy about reading in the blobs.
             int blobCount = deserializer.ReadInt();
             SegmentedMemoryStreamWriter writer = new SegmentedMemoryStreamWriter(blobCount);
             while (8 <= blobCount)
@@ -602,17 +602,17 @@ namespace Graphs
 
             m_reader = writer.GetReader();
 
-            // Stuff added in version 1.   See Version below 
+            // Stuff added in version 1.   See Version below
             if (1 <= deserializer.MinimumReaderVersionBeingRead)
             {
                 // Because Graph has superclass, you can't add objects to the end of it (since it is not 'the end' of the object)
                 // which is a problem if we want to add new fields.  We could have had a worker object but another way of doing
                 // it is create a deferred (lazy region).   The key is that ALL readers know how to skip this region, which allows
-                // you to add new fields 'at the end' of the region (just like for sealed objects).  
+                // you to add new fields 'at the end' of the region (just like for sealed objects).
                 DeferedRegion expansion = new DeferedRegion();
                 expansion.Read(deserializer, delegate
                 {
-                    // I don't need to use Tagged types for my 'first' version of this new region 
+                    // I don't need to use Tagged types for my 'first' version of this new region
                     int count = deserializer.ReadInt();
                     for (int i = 0; i < count; i++)
                     {
@@ -625,30 +625,30 @@ namespace Graphs
                     }
 
                     // You can add any tagged objects here after version 1.   You can also use the deserializer.VersionBeingRead
-                    // to avoid reading non-existent fields, but the tagging is probably better.   
+                    // to avoid reading non-existent fields, but the tagging is probably better.
                 });
-                expansion.FinishRead(true);  // Immediately read in the fields, preserving the current position in the stream.     
+                expansion.FinishRead(true);  // Immediately read in the fields, preserving the current position in the stream.
             }
         }
 
-        // These three members control the versioning of the Graph format on disk.   
-        public int Version { get { return 1; } }                            // The version of what was written.  It is in the file.       
+        // These three members control the versioning of the Graph format on disk.
+        public int Version { get { return 1; } }                            // The version of what was written.  It is in the file.
         public int MinimumVersionCanRead { get { return 0; } }              // Declaration of the oldest format this code can read
-        public int MinimumReaderVersion                                     // Will cause readers to fail if their code version is less than this.  
+        public int MinimumReaderVersion                                     // Will cause readers to fail if their code version is less than this.
         {
             get
             {
                 if (m_deferedTypes.Count != 0)
                 {
-                    return 1;    // We require that you upgrade to version 1 if you use m_deferedTypes (e.g. projectN)   
+                    return 1;    // We require that you upgrade to version 1 if you use m_deferedTypes (e.g. projectN)
                 }
 
                 return 0;
             }
         }
 
-        private int m_expectedNodeCount;                // Initial guess at graph Size. 
-        private long m_totalSize;                       // Total Size of all the nodes in the graph.  
+        private int m_expectedNodeCount;                // Initial guess at graph Size.
+        private long m_totalSize;                       // Total Size of all the nodes in the graph.
         internal int m_totalRefs;                       // Total Number of references in the graph
         internal GrowableArray<TypeInfo> m_types;       // We expect only thousands of these
         internal GrowableArray<DeferedTypeInfo> m_deferedTypes; // Types that we only have IDs and module image bases.
@@ -663,11 +663,11 @@ namespace Graphs
     }
 
     /// <summary>
-    /// Node represents a single node in the code:Graph.  These are created lazily and follow a pattern were the 
+    /// Node represents a single node in the code:Graph.  These are created lazily and follow a pattern were the
     /// CALLER provides the storage for any code:Node or code:NodeType value that are returned.   Thus the caller
-    /// is responsible for determine when nodes can be reused to minimize GC cost.  
-    /// 
-    /// A node implicitly knows where the 'next' child is (that is it is an iterator).  
+    /// is responsible for determine when nodes can be reused to minimize GC cost.
+    ///
+    /// A node implicitly knows where the 'next' child is (that is it is an iterator).
     /// </summary>
     [ExcludeFromCodeCoverage]
     internal class Node
@@ -695,7 +695,7 @@ namespace Graphs
         }
 
         /// <summary>
-        /// Reset the internal state so that 'GetNextChildIndex; will return the first child.  
+        /// Reset the internal state so that 'GetNextChildIndex; will return the first child.
         /// </summary>
         public void ResetChildrenEnumeration()
         {
@@ -711,7 +711,7 @@ namespace Graphs
         }
 
         /// <summary>
-        /// Gets the index of the first child of node.  Will return NodeIndex.Invalid if there are no children. 
+        /// Gets the index of the first child of node.  Will return NodeIndex.Invalid if there are no children.
         /// </summary>
         /// <returns>The index of the child </returns>
         public NodeIndex GetFirstChildIndex()
@@ -737,7 +737,7 @@ namespace Graphs
         }
 
         /// <summary>
-        /// Returns the number of children this node has.  
+        /// Returns the number of children this node has.
         /// </summary>
         public int ChildCount
         {
@@ -764,7 +764,7 @@ namespace Graphs
         public NodeIndex Index { get { return m_index; } }
         public Graph Graph { get { return m_graph; } }
         /// <summary>
-        /// Returns true if 'node' is a child of 'this'.  childStorage is simply used as temp space 
+        /// Returns true if 'node' is a child of 'this'.  childStorage is simply used as temp space
         /// as was allocated by Graph.AllocateNodeStorage
         /// </summary>
         public bool Contains(NodeIndex nodeIndex)
@@ -839,7 +839,7 @@ namespace Graphs
             m_index = NodeIndex.Invalid;
         }
 
-        // Node information is stored in a compressed form because we have a lot of them. 
+        // Node information is stored in a compressed form because we have a lot of them.
         internal static int ReadCompressedInt(SegmentedMemoryStreamReader reader)
         {
             int ret = 0;
@@ -922,13 +922,13 @@ namespace Graphs
     }
 
     /// <summary>
-    /// Represents the nodeId of a particular node in the graph.  
+    /// Represents the nodeId of a particular node in the graph.
     /// </summary>
     [ExcludeFromCodeCoverage]
     internal sealed class NodeType
     {
         /// <summary>
-        /// Every nodeId has a name, this is it.  
+        /// Every nodeId has a name, this is it.
         /// </summary>
         public string Name
         {
@@ -954,7 +954,7 @@ namespace Graphs
             }
         }
         /// <summary>
-        /// This is the ModuleName ! Name (or just Name if ModuleName does not exist)  
+        /// This is the ModuleName ! Name (or just Name if ModuleName does not exist)
         /// </summary>
         public string FullName
         {
@@ -966,7 +966,7 @@ namespace Graphs
                     return Name;
                 }
 
-                if (moduleName.Length == 0) // TODO should we have this convention?   
+                if (moduleName.Length == 0) // TODO should we have this convention?
                 {
                     moduleName = "?";
                 }
@@ -975,17 +975,17 @@ namespace Graphs
             }
         }
         /// <summary>
-        /// Size is defined as the Size of the first node in the graph of a given nodeId.   
+        /// Size is defined as the Size of the first node in the graph of a given nodeId.
         /// For types that always have the same Size this is useful, but for types (like arrays or strings)
-        /// that have variable Size, it is not useful.  
-        /// 
+        /// that have variable Size, it is not useful.
+        ///
         /// TODO keep track if the nodeId is of variable Size
         /// </summary>
         public int Size { get { return m_graph.m_types[(int)m_index].Size; } }
         public NodeTypeIndex Index { get { return m_index; } }
         public Graph Graph { get { return m_graph; } }
         /// <summary>
-        /// The module associated with the type.  Can be null.  Typically this is the full path name.  
+        /// The module associated with the type.  Can be null.  Typically this is the full path name.
         /// </summary>
         public string ModuleName
         {
@@ -1058,7 +1058,7 @@ namespace Graphs
         /// </summary>
         public int Size;
         /// <summary>
-        /// The time when this image was built (There is a field in the PE header).   May be MinimumValue if unknown. 
+        /// The time when this image was built (There is a field in the PE header).   May be MinimumValue if unknown.
         /// </summary>
         public DateTime BuildTime;      // From in the PE header
         /// <summary>
@@ -1066,17 +1066,17 @@ namespace Graphs
         /// </summary>
         public string PdbName;
         /// <summary>
-        /// The GUID that uniquely identifies this PDB for symbol server lookup.  May be Guid.Empty it not known.  
+        /// The GUID that uniquely identifies this PDB for symbol server lookup.  May be Guid.Empty it not known.
         /// </summary>
-        public Guid PdbGuid;            // PDB Guid 
+        public Guid PdbGuid;            // PDB Guid
         /// <summary>
-        /// The age (version number) that is used for symbol server lookup.  
+        /// The age (version number) that is used for symbol server lookup.
         /// </summary>T
         public int PdbAge;
 
         #region private
         /// <summary>
-        /// Implementing IFastSerializable interface.  
+        /// Implementing IFastSerializable interface.
         /// </summary>
         public void ToStream(Serializer serializer)
         {
@@ -1089,7 +1089,7 @@ namespace Graphs
             serializer.Write(PdbAge);
         }
         /// <summary>
-        /// Implementing IFastSerializable interface.  
+        /// Implementing IFastSerializable interface.
         /// </summary>
         public void FromStream(Deserializer deserializer)
         {
@@ -1105,28 +1105,28 @@ namespace Graphs
     }
 
     /// <summary>
-    /// Each node is given a unique index (which is dense: an array is a good lookup structure).   
+    /// Each node is given a unique index (which is dense: an array is a good lookup structure).
     /// To avoid passing the wrong indexes to methods, we make an enum for each index.   This does
     /// mean you need to cast away this strong typing occasionally (e.g. when you index arrays)
-    /// However on the whole it is a good tradeoff.  
+    /// However on the whole it is a good tradeoff.
     /// </summary>
     public enum NodeIndex { Invalid = -1 }
     /// <summary>
-    /// Each node nodeId is given a unique index (which is dense: an array is a good lookup structure).   
+    /// Each node nodeId is given a unique index (which is dense: an array is a good lookup structure).
     /// To avoid passing the wrong indexes to methods, we make an enum for each index.   This does
     /// mean you need to cast away this strong typing occasionally (e.g. when you index arrays)
-    /// However on the whole it is a good tradeoff.  
-    /// </summary>    
+    /// However on the whole it is a good tradeoff.
+    /// </summary>
     public enum NodeTypeIndex { Invalid = -1 }
 
     /// <summary>
-    /// Stuff that is useful but does not need to be in Graph.   
+    /// Stuff that is useful but does not need to be in Graph.
     /// </summary>
     [ExcludeFromCodeCoverage]
     internal static class GraphUtils
     {
         /// <summary>
-        /// Write the graph as XML to a string and return it (useful for debugging small graphs).  
+        /// Write the graph as XML to a string and return it (useful for debugging small graphs).
         /// </summary>
         /// <returns></returns>
         public static string PrintGraph(this Graph graph)
@@ -1166,9 +1166,9 @@ namespace Graphs
         {
             return graph.PrintChildren((NodeIndex)nodeIndex);
         }
-        // Debugging. 
+        // Debugging.
         /// <summary>
-        /// Writes the graph as XML to 'writer'.  Don't use on big graphs.  
+        /// Writes the graph as XML to 'writer'.  Don't use on big graphs.
         /// </summary>
         public static void WriteXml(this Graph graph, TextWriter writer)
         {
@@ -1200,7 +1200,7 @@ namespace Graphs
             NodeType typeStorage = graph.AllocTypeNodeStorage();
             Node node;
 
-#if false 
+#if false
             // Compute reachability info
             bool[] reachable = new bool[(int)graph.NodeIndexLimit];
             Queue<NodeIndex> workQueue = new Queue<NodeIndex>();
@@ -1217,14 +1217,14 @@ namespace Graphs
                 }
             }
 
-            // Get Reachability count. 
+            // Get Reachability count.
             int reachableCount = 0;
             for (int i = 0; i < reachable.Length; i++)
                 if (reachable[i])
                     reachableCount++;
 #endif
 
-            // Sort the nodes by virtual address 
+            // Sort the nodes by virtual address
             NodeIndex[] sortedNodes = new NodeIndex[(int)graph.NodeIndexLimit];
             for (int i = 0; i < sortedNodes.Length; i++)
             {
@@ -1329,1332 +1329,12 @@ namespace Graphs
     }
 }
 
-/// <summary>
-/// A RefGraph is derived graph where each node's children are the set of nodes in the original graph 
-/// which refer that node (that is A -> B then in refGraph B -> A).   
-/// 
-/// The NodeIndexes in the refGraph match the NodeIndexes in the original graph.  Thus after creating
-/// a refGraph it is easy to answer the question 'who points at me' of the original graph.  
-/// 
-/// When create the RefGraph the whole reference graph is generated on the spot (thus it must traverse
-/// the whole of the original graph) and the size of the resulting RefGraph is  about the same size as the  
-/// original graph. 
-/// 
-/// Thus this is a fairly expensive thing to create.  
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class RefGraph
-{
-    public RefGraph(Graph graph)
-    {
-        m_refsForNodes = new NodeListIndex[(int)graph.NodeIndexLimit];
-        // We guess that we need about 1.5X as many slots as there are nodes.   This seems a conservative estimate. 
-        m_links = new GrowableArray<RefElem>((int)graph.NodeIndexLimit * 3 / 2);
-
-        var nodeStorage = graph.AllocNodeStorage();
-        for (NodeIndex nodeIndex = 0; nodeIndex < graph.NodeIndexLimit; nodeIndex++)
-        {
-            var node = graph.GetNode(nodeIndex, nodeStorage);
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                AddRefsTo(childIndex, nodeIndex);
-            }
-        }
-
-        // Sadly, this check is too expensive even for DEBUG 
-#if false 
-        CheckConsistancy(graph);
-#endif
-    }
-    /// <summary>
-    /// Allocates nodes to be used as storage for methods like code:GetNode, code:RefNode.GetFirstChild and code:RefNode.GetNextChild
-    /// </summary>
-    public RefNode AllocNodeStorage() { return new RefNode(this); }
-
-    /// <summary>
-    /// Given an arbitrary code:NodeIndex that identifies the node, Get a code:Node object.  
-    /// 
-    /// This routine does not allocated but uses the space passed in by 'storage.  
-    /// 'storage' should be allocated with code:AllocNodeStorage, and should be aggressively reused.  
-    /// </summary>
-    public RefNode GetNode(NodeIndex nodeIndex, RefNode storage)
-    {
-        Debug.Assert(storage.m_graph == this);
-        storage.m_index = nodeIndex;
-        return storage;
-    }
-
-    /// <summary>
-    /// This is for debugging 
-    /// </summary>
-    /// <param name="nodeIndex"></param>
-    /// <returns></returns>
-    public RefNode GetNode(NodeIndex nodeIndex)
-    {
-        return GetNode(nodeIndex, AllocNodeStorage());
-    }
-
-    #region private
-#if DEBUG
-    private void CheckConsitancy(Graph graph)
-    {
-        // This double check is pretty expensive for large graphs (nodes that have large fan-in or fan-out).  
-        var nodeStorage = graph.AllocNodeStorage();
-        var refStorage = AllocNodeStorage();
-        for (NodeIndex nodeIdx = 0; nodeIdx < graph.NodeIndexLimit; nodeIdx++)
-        {
-            // If Node -> Ref then the RefGraph has a pointer from Ref -> Node 
-            var node = graph.GetNode(nodeIdx, nodeStorage);
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                var refsForChild = GetNode(childIndex, refStorage);
-                if (!refsForChild.Contains(nodeIdx))
-                {
-                    _ = node.ToString();
-                    _ = refsForChild.ToString();
-                    Debug.Assert(false);
-                }
-            }
-
-            // If the refs graph has a pointer from Ref -> Node then the original graph has a arc from Node ->Ref
-            var refNode = GetNode(nodeIdx, refStorage);
-            for (var childIndex = refNode.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = refNode.GetNextChildIndex())
-            {
-                var nodeForChild = graph.GetNode(childIndex, nodeStorage);
-                if (!nodeForChild.Contains(nodeIdx))
-                {
-                    _ = nodeForChild.ToString();
-                    _ = refNode.ToString();
-                    Debug.Assert(false);
-                }
-            }
-        }
-    }
-#endif
-
-    /// <summary>
-    /// Add the fact that 'refSource' refers to refTarget.
-    /// </summary>
-    private void AddRefsTo(NodeIndex refTarget, NodeIndex refSource)
-    {
-        NodeListIndex refsToList = m_refsForNodes[(int)refTarget];
-
-        // We represent singles as the childIndex itself.  This is a very common case, so it is good that it is efficient. 
-        if (refsToList == NodeListIndex.Empty)
-        {
-            m_refsForNodes[(int)refTarget] = (NodeListIndex)(refSource + 1);
-        }
-        else if (refsToList > 0)        // One element list
-        {
-            var existingChild = (NodeIndex)(refsToList - 1);
-            m_refsForNodes[(int)refTarget] = (NodeListIndex)(-AddLink(refSource, AddLink(existingChild)) - 1);
-        }
-        else // refsToList < 0          more than one element.  
-        {
-            var listIndex = -(int)refsToList - 1;
-            m_refsForNodes[(int)refTarget] = (NodeListIndex)(-AddLink(refSource, listIndex) - 1);
-        }
-    }
-
-    /// <summary>
-    /// A helper function for AddRefsTo.  Allocates a new cell from m_links and initializes its two fields 
-    /// (the child index field and 'rest' field), and returns the index (pointer) to the new cell.  
-    /// </summary>
-    private int AddLink(NodeIndex refIdx, int nextIdx = -1)
-    {
-        var ret = m_links.Count;
-        m_links.Add(new RefElem(refIdx, nextIdx));
-        return ret;
-    }
-
-    /// <summary>
-    ///  Logically a NodeListIndex represents a list of node indexes.   However it is heavily optimized
-    ///  to avoid overhead.   0 means empty, a positive number is the NodeIndex+1 and a negative number 
-    ///  is index in m_links - 1.  
-    /// </summary>
-    internal enum NodeListIndex { Empty = 0 };
-
-    /// <summary>
-    /// RefElem is a linked list cell that is used to store lists of children that are larger than 1.
-    /// </summary>
-    internal struct RefElem
-    {
-        public RefElem(NodeIndex refIdx, int nextIdx) { RefIdx = refIdx; NextIdx = nextIdx; }
-        public NodeIndex RefIdx;           // The reference
-        public int NextIdx;                // The index to the next element in  m_links.   a negative number when done. 
-    }
-
-    /// <summary>
-    /// m_refsForNodes maps the NodeIndexes of the reference graph to the children information.   However unlike
-    /// a normal Graph RefGraph needs to support incremental addition of children.  Thus we can't use the normal
-    /// compression (which assumed you know all the children when you define the node).  
-    /// 
-    /// m_refsForNodes points at a NodeListIndex which is a compressed list that is tuned for the case where
-    /// a node has exactly one child (a very common case).   If that is not true we 'overflow' into a 'linked list'
-    /// of RefElems that is stored in m_links.   See NodeListIndex for more on the exact encoding.   
-    /// 
-    /// </summary>
-    internal NodeListIndex[] m_refsForNodes;
-
-    /// <summary>
-    /// If the number of children for a node is > 1 then we need to store the data somewhere.  m_links is array
-    /// of linked list cells that hold the overflow case.  
-    /// </summary>
-    internal GrowableArray<RefElem> m_links;      // The rest of the list.  
-    #endregion
-}
-
-[ExcludeFromCodeCoverage]
-internal sealed class RefNode
-{
-    /// <summary>
-    /// Gets the first child for the node.  Will return null if there are no children.  
-    /// </summary>
-    public NodeIndex GetFirstChildIndex()
-    {
-        var refsToList = m_graph.m_refsForNodes[(int)m_index];
-
-        if (refsToList == RefGraph.NodeListIndex.Empty)
-        {
-            return NodeIndex.Invalid;
-        }
-
-        if (refsToList > 0)        // One element list
-        {
-            m_cur = -1;
-            return (NodeIndex)(refsToList - 1);
-        }
-        else // refsToList < 0          more than one element.  
-        {
-            var listIndex = -(int)refsToList - 1;
-            var refElem = m_graph.m_links[listIndex];
-            m_cur = refElem.NextIdx;
-            return refElem.RefIdx;
-        }
-    }
-    /// <summary>
-    /// Returns the next child for the node.   Will return NodeIndex.Invalid if there are no more children 
-    /// </summary>
-    public NodeIndex GetNextChildIndex()
-    {
-        if (m_cur < 0)
-        {
-            return NodeIndex.Invalid;
-        }
-
-        var refElem = m_graph.m_links[m_cur];
-        m_cur = refElem.NextIdx;
-        return refElem.RefIdx;
-    }
-
-    /// <summary>
-    /// Returns the count of children (nodes that reference this node). 
-    /// </summary>
-    public int ChildCount
-    {
-        get
-        {
-            var ret = 0;
-            for (NodeIndex childIndex = GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = GetNextChildIndex())
-            {
-                ret++;
-            }
-
-            return ret;
-        }
-    }
-
-    public RefGraph Graph { get { return m_graph; } }
-    public NodeIndex Index { get { return m_index; } }
-
-    /// <summary>
-    /// Returns true if 'node' is a child of 'this'.  childStorage is simply used as temp space 
-    /// as was allocated by RefGraph.AllocateNodeStorage
-    /// </summary>
-    public bool Contains(NodeIndex node)
-    {
-        for (NodeIndex childIndex = GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = GetNextChildIndex())
-        {
-            if (childIndex == node)
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public override string ToString()
-    {
-        StringWriter sw = new StringWriter();
-        WriteXml(sw);
-        return sw.ToString();
-    }
-    public void WriteXml(TextWriter writer, string prefix = "")
-    {
-        Debug.Assert(Index != NodeIndex.Invalid);
-
-
-        writer.Write("{0}<Node Index=\"{1}\" NumChildren=\"{2}\"", prefix, (int)Index, ChildCount);
-        var childIndex = GetFirstChildIndex();
-        if (childIndex != NodeIndex.Invalid)
-        {
-            writer.WriteLine(">");
-            writer.Write(prefix);
-            int i = 0;
-            do
-            {
-                writer.Write(" {0}", childIndex);
-                childIndex = GetNextChildIndex();
-                i++;
-                if (i >= 32)
-                {
-                    writer.WriteLine();
-                    writer.Write(prefix);
-                    i = 0;
-                }
-            } while (childIndex != NodeIndex.Invalid);
-            writer.WriteLine(" </Node>");
-        }
-        else
-        {
-            writer.WriteLine("/>");
-        }
-    }
-
-    #region private
-    internal RefNode(RefGraph refGraph)
-    {
-        m_graph = refGraph;
-    }
-
-    internal RefGraph m_graph;
-    internal NodeIndex m_index;     // My index.  
-    internal int m_cur;             // A pointer to where we are in the list of elements (index into m_links)
-    #endregion
-}
-
-/// <summary>
-/// code:MemorySampleSource hooks up a Memory graph to become a Sample source.  Currently we do
-/// a breadth-first traversal to form a spanning tree, and then create samples for each node
-/// where the 'stack' is the path to the root of this spanning tree.
-/// 
-/// This is just a first cut...
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class SpanningTree
-{
-    public SpanningTree(Graph graph, TextWriter log)
-    {
-        Graph = graph;
-        m_log = log;
-        m_nodeStorage = graph.AllocNodeStorage();
-        m_childStorage = graph.AllocNodeStorage();
-        m_typeStorage = graph.AllocTypeNodeStorage();
-
-        // We need to reduce the graph to a tree.   Each node is assigned a unique 'parent' which is its 
-        // parent in a spanning tree of the graph.  
-        // The +1 is for orphan node support.  
-        m_parent = new NodeIndex[(int)graph.NodeIndexLimit + 1];
-    }
-    public Graph Graph { get; }
-
-    /// <summary>
-    /// Every type is given a priority of 0 unless the type name matches one of 
-    /// the patterns in PriorityRegExs.  If it does that type is assigned that priority.
-    /// 
-    /// A node's priority is defined to be the priority of the type of the node
-    /// (as given by PriorityRegExs), plus 1/10 the priority of its parent.  
-    /// 
-    /// Thus priorities 'decay' by 1/10 through pointers IF the priority of the node's
-    /// type is 0 (the default).   
-    ///
-    /// By default the framework has a priority of -1 which means that you explore all
-    /// high priority and user defined types before any framework type.
-    /// 
-    /// Types with the same priority are enumerate breath-first.  
-    /// </summary>
-    public string PriorityRegExs
-    {
-        get
-        {
-            if (m_priorityRegExs == null)
-            {
-                PriorityRegExs = DefaultPriorities;
-            }
-
-            return m_priorityRegExs;
-        }
-        set
-        {
-            m_priorityRegExs = value;
-            SetTypePriorities(value);
-        }
-    }
-    public static string DefaultPriorities
-    {
-        get
-        {
-            return
-                // By types (including user defined types) are 0
-                @"v4.0.30319\%!->-1;" +     // Framework is less than default
-                @"v2.0.50727\%!->-1;" +     // Framework is less than default
-                @"[*local vars]->-1000;" +  // Local variables are not that interesting, since they tend to be transient
-                @"mscorlib!Runtime.CompilerServices.ConditionalWeakTable->-10000;" + // We prefer not to use Conditional weak table references even more. 
-                @"[COM/WinRT Objects]->-1000000;" + // We prefer to Not use the CCW roots. 
-                @"[*handles]->-2000000;" +
-                @"[other roots]->-2000000";
-        }
-    }
-
-    public NodeIndex Parent(NodeIndex node) { return m_parent[(int)node]; }
-
-    public void ForEach(Action<NodeIndex> callback)
-    {
-        // Initialize the priority 
-        if (m_typePriorities == null)
-        {
-            PriorityRegExs = DefaultPriorities;
-        }
-
-        Debug.Assert(m_typePriorities != null);
-
-        // Initialize the breadth-first work queue.
-        var nodesToVisit = new PriorityQueue(1024);
-        nodesToVisit.Enqueue(Graph.RootIndex, 0.0F);
-
-        // reset the visited information.
-        for (int i = 0; i < m_parent.Length; i++)
-        {
-            m_parent[i] = NodeIndex.Invalid;
-        }
-
-        float[] nodePriorities = new float[m_parent.Length];
-        bool scanedForOrphans = false;
-        var epsilon = 1E-7F;            // Something that is big enough not to bet lost in roundoff error.  
-        float order = 0;
-        for (int i = 0; ; i++)
-        {
-            if ((i & 0x1FFF) == 0)  // Every 8K
-            {
-                System.Threading.Thread.Sleep(0);       // Allow interruption.  
-            }
-
-            NodeIndex nodeIndex;
-            float nodePriority;
-            if (nodesToVisit.Count == 0)
-            {
-                nodePriority = 0;
-                if (!scanedForOrphans)
-                {
-                    scanedForOrphans = true;
-                    AddOrphansToQueue(nodesToVisit);
-                }
-                if (nodesToVisit.Count == 0)
-                {
-                    return;
-                }
-            }
-            nodeIndex = nodesToVisit.Dequeue(out nodePriority);
-
-            // Insert any children that have not already been visited (had a parent assigned) into the work queue). 
-            var node = Graph.GetNode(nodeIndex, m_nodeStorage);
-            var parentPriority = nodePriorities[(int)node.Index];
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                if (m_parent[(int)childIndex] == NodeIndex.Invalid)
-                {
-                    m_parent[(int)childIndex] = nodeIndex;
-
-                    // the priority of the child is determined by its type and 1/10 by its parent.  
-                    var child = Graph.GetNode(childIndex, m_childStorage);
-                    var childPriority = m_typePriorities[(int)child.TypeIndex] + parentPriority / 10;
-                    nodePriorities[(int)childIndex] = childPriority;
-
-                    // Subtract a small increasing value to keep the queue in order if the priorities are the same. 
-                    // This is a bit of a hack since it can get big and purtub the user-defined order.  
-                    order += epsilon;
-                    nodesToVisit.Enqueue(childIndex, childPriority - order);
-                }
-            }
-
-            // Return the node.  
-            callback?.Invoke(node.Index);
-        }
-    }
-
-    #region private
-    /// <summary>
-    /// Add any unreachable nodes to the 'nodesToVisit'.   Note that we do this in a 'smart' way
-    /// where we only add orphans that are not reachable from other orphans.   That way we get a 
-    /// minimal set of orphan 'roots'.  
-    /// </summary>
-    /// <param name="nodesToVisit"></param>
-    private void AddOrphansToQueue(PriorityQueue nodesToVisit)
-    {
-
-        for (int i = 0; i < (int)Graph.NodeIndexLimit; i++)
-        {
-            if (m_parent[i] == NodeIndex.Invalid)
-            {
-                MarkDecendentsIgnoringCycles((NodeIndex)i, 0);
-            }
-        }
-
-        // Collect up all the nodes that are not reachable from other nodes as the roots of the
-        // orphans.  Also reset orphanVisitedMarker back to NodeIndex.Invalid.
-        for (int i = 0; i < (int)Graph.NodeIndexLimit; i++)
-        {
-            var nodeIndex = (NodeIndex)i;
-            var parent = m_parent[(int)nodeIndex];
-            if (parent <= NodeIndex.Invalid)
-            {
-                if (parent == NodeIndex.Invalid)
-                {
-                    // Thr root index has no parent but is reachable from the root. 
-                    if (nodeIndex != Graph.RootIndex)
-                    {
-                        var node = Graph.GetNode(nodeIndex, m_nodeStorage);
-                        var priority = m_typePriorities[(int)node.TypeIndex];
-                        nodesToVisit.Enqueue(nodeIndex, priority);
-                        m_parent[(int)nodeIndex] = Graph.NodeIndexLimit;               // This is the 'not reachable' parent. 
-                    }
-                }
-                else
-                {
-                    m_parent[(int)nodeIndex] = NodeIndex.Invalid;
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// A helper for AddOrphansToQueue, so we only add orphans that are not reachable from other orphans.  
-    /// 
-    /// Mark all descendants (but not nodeIndex itself) as being visited.    Any arcs that form
-    /// cycles are ignored, so nodeIndex is guaranteed to NOT be marked.     
-    /// </summary>
-    private void MarkDecendentsIgnoringCycles(NodeIndex nodeIndex, int recursionCount)
-    {
-        // TODO We give up if the chains are larger than 10K long (because we stack overflow otherwise)
-        // We could have an explicit stack and avoid this...
-        if (recursionCount > 10000)
-        {
-            return;
-        }
-
-        Debug.Assert(m_parent[(int)nodeIndex] == NodeIndex.Invalid);
-
-        // This marks that there is a path from another orphan to this one (thus it is not a good root)
-        NodeIndex orphanVisitedMarker = NodeIndex.Invalid - 1;
-
-        // To detect cycles we mark all nodes we not committed to (we are visiting, rather than visited)
-        // If we detect this mark we understand it is a loop and ignore the arc.  
-        NodeIndex orphanVisitingMarker = NodeIndex.Invalid - 2;
-        m_parent[(int)nodeIndex] = orphanVisitingMarker;        // We are now visitING
-
-        // Mark all nodes as being visited.  
-        var node = Graph.GetNode(nodeIndex, AllocNodeStorage());
-        for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-        {
-            // Has this child not been seen at all?  If so mark it.  
-            // Skip it if we are visiting (it would form a cycle) or visited (or not an orphan)
-            if (m_parent[(int)childIndex] == NodeIndex.Invalid)
-            {
-                MarkDecendentsIgnoringCycles(childIndex, recursionCount + 1);
-                m_parent[(int)childIndex] = orphanVisitedMarker;
-            }
-        }
-        FreeNodeStorage(node);
-
-        // We set this above, and should not have changed it.  
-        Debug.Assert(m_parent[(int)nodeIndex] == orphanVisitingMarker);
-        // Now that we are finished, we reset the visiting bit.  
-        m_parent[(int)nodeIndex] = NodeIndex.Invalid;
-    }
-
-    /// <summary>
-    /// Gives back nodes that are no longer in use.  This is a memory optimization. 
-    /// </summary>
-    private void FreeNodeStorage(Node node)
-    {
-        m_cachedNodeStorage = node;
-    }
-    /// <summary>
-    /// Gets a node that can be written on.  It is a simple cache
-    /// </summary>
-    /// <returns></returns>
-    private Node AllocNodeStorage()
-    {
-        var ret = m_cachedNodeStorage;                // See if we have a free node. 
-        if (ret == null)
-        {
-            ret = Graph.AllocNodeStorage();
-        }
-        else
-        {
-            m_cachedNodeStorage = null;               // mark that that node is in use.  
-        }
-
-        return ret;
-    }
-
-    /// <summary>
-    /// Convert a string from my regular expression format (where you only have * and {  } as grouping operators
-    /// and convert them to .NET regular expressions string
-    /// TODO FIX NOW cloned code (also in FilterStackSource)
-    /// </summary>
-    internal static string ToDotNetRegEx(string str)
-    {
-        // A leading @ sign means the rest is a .NET regular expression.  (Undocumented, not really needed yet.)
-        if (str.StartsWith('@'))
-        {
-            return str.Substring(1);
-        }
-
-        str = Regex.Escape(str);                // Assume everything is ordinary
-        str = str.Replace(@"%", @"[.\w\d?]*");  // % means any number of alpha-numeric chars. 
-        str = str.Replace(@"\*", @".*");        // * means any number of any characters.  
-        str = str.Replace(@"\^", @"^");         // ^ means anchor at the beginning.  
-        str = str.Replace(@"\|", @"|");         // | means is the or operator  
-        str = str.Replace(@"\{", "(");
-        str = str.Replace('}', ')');
-        return str;
-    }
-
-    private void SetTypePriorities(string priorityPats)
-    {
-        m_typePriorities ??= new float[(int)Graph.NodeTypeIndexLimit];
-
-        string[] priorityPatArray = priorityPats.Split(';');
-        Regex[] priorityRegExArray = new Regex[priorityPatArray.Length];
-        float[] priorityArray = new float[priorityPatArray.Length];
-        for (int i = 0; i < priorityPatArray.Length; i++)
-        {
-            var m = Regex.Match(priorityPatArray[i], @"(.*)->(-?\d+.?\d*)");
-            if (!m.Success)
-            {
-                if (string.IsNullOrWhiteSpace(priorityPatArray[i]))
-                {
-                    continue;
-                }
-
-                throw new ApplicationException($"Priority pattern {priorityPatArray[i]} is not of the form Pat->Num.");
-            }
-
-            var dotNetRegEx = ToDotNetRegEx(m.Groups[1].Value.Trim());
-            priorityRegExArray[i] = new Regex(dotNetRegEx, RegexOptions.IgnoreCase);
-            priorityArray[i] = float.Parse(m.Groups[2].Value);
-        }
-
-        // Assign every type index a priority in m_typePriorities based on if they match a pattern.  
-        NodeType typeStorage = Graph.AllocTypeNodeStorage();
-        for (NodeTypeIndex typeIdx = 0; typeIdx < Graph.NodeTypeIndexLimit; typeIdx++)
-        {
-            var type = Graph.GetType(typeIdx, typeStorage);
-
-            var fullName = type.FullName;
-            for (int regExIdx = 0; regExIdx < priorityRegExArray.Length; regExIdx++)
-            {
-                var priorityRegEx = priorityRegExArray[regExIdx];
-                if (priorityRegEx == null)
-                {
-                    continue;
-                }
-
-                var m = priorityRegEx.Match(fullName);
-                if (m.Success)
-                {
-                    m_typePriorities[(int)typeIdx] = priorityArray[regExIdx];
-                    // m_log.WriteLine("Type {0} assigned priority {1:f3}", fullName, priorityArray[regExIdx]);
-                    break;
-                }
-            }
-        }
-    }
-
-    private NodeIndex[] m_parent;               // We keep track of the parents of each node in our breadth-first scan. 
-
-    // We give each type a priority (using the m_priority Regular expressions) which guide the breadth-first scan. 
-    private string m_priorityRegExs;
-    private float[] m_typePriorities;
-    private NodeType m_typeStorage;
-    private Node m_nodeStorage;                 // Only for things that can't be reentrant
-    private Node m_childStorage;
-    private Node m_cachedNodeStorage;           // Used when it could be reentrant
-    private TextWriter m_log;                   // processing messages 
-    #endregion
-}
-
-/// <summary>
-/// TODO FIX NOW put in its own file.  
-/// A priority queue, specialized to be a bit more efficient than a generic version would be. 
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class PriorityQueue
-{
-    public PriorityQueue(int initialSize = 32)
-    {
-        m_heap = new DataItem[initialSize];
-    }
-    public int Count { get; private set; }
-
-    public void Enqueue(NodeIndex item, float priority)
-    {
-        var idx = Count;
-        if (idx >= m_heap.Length)
-        {
-            var newArray = new DataItem[m_heap.Length * 3 / 2 + 8];
-            Array.Copy(m_heap, newArray, m_heap.Length);
-            m_heap = newArray;
-        }
-        m_heap[idx].value = item;
-        m_heap[idx].priority = priority;
-        Count = idx + 1;
-        for (; ; )
-        {
-            var parent = idx / 2;
-            if (m_heap[parent].priority >= m_heap[idx].priority)
-            {
-                break;
-            }
-
-            (m_heap[idx], m_heap[parent]) = (m_heap[parent], m_heap[idx]);
-
-            if (parent == 0)
-            {
-                break;
-            }
-
-            idx = parent;
-        }
-        // CheckInvariant();
-    }
-    public NodeIndex Dequeue(out float priority)
-    {
-        Debug.Assert(Count > 0);
-
-        var ret = m_heap[0].value;
-        priority = m_heap[0].priority;
-        --Count;
-        m_heap[0] = m_heap[Count];
-        var idx = 0;
-        for (; ; )
-        {
-            var childIdx = idx * 2;
-            var largestIdx = idx;
-            if (childIdx < Count && m_heap[childIdx].priority > m_heap[largestIdx].priority)
-            {
-                largestIdx = childIdx;
-            }
-
-            childIdx++;
-            if (childIdx < Count && m_heap[childIdx].priority > m_heap[largestIdx].priority)
-            {
-                largestIdx = childIdx;
-            }
-
-            if (largestIdx == idx)
-            {
-                break;
-            }
-
-            (m_heap[idx], m_heap[largestIdx]) = (m_heap[largestIdx], m_heap[idx]);
-
-            idx = largestIdx;
-        }
-        // CheckInvariant();
-        return ret;
-    }
-
-    #region private
-#if DEBUG
-    public override string ToString()
-    {
-        var sb = new StringBuilder();
-        sb.AppendLine("<PriorityQueue Count=\"").Append(Count).Append("\">").AppendLine();
-
-        // Sort the items in descending order 
-        var items = new List<DataItem>(Count);
-        for (int i = 0; i < Count; i++)
-            items.Add(m_heap[i]);
-        items.Sort((x, y) => y.priority.CompareTo(x.priority));
-        if (items.Count > 0)
-            Debug.Assert(items[0].value == m_heap[0].value);
-
-        foreach (var item in items)
-            sb.Append("{").Append((int)item.value).Append(", ").Append(item.priority.ToString("f1")).Append("}").AppendLine();
-        sb.AppendLine("</PriorityQueue>");
-        return sb.ToString();
-    }
-#endif
-
-    private struct DataItem
-    {
-        public DataItem(NodeIndex value, float priority) { this.value = value; this.priority = priority; }
-        public float priority;
-        public NodeIndex value;
-    }
-    [Conditional("DEBUG")]
-    private void CheckInvariant()
-    {
-        for (int idx = 1; idx < Count; idx++)
-        {
-            var parentIdx = idx / 2;
-            Debug.Assert(m_heap[parentIdx].priority >= m_heap[idx].priority);
-        }
-    }
-
-    // In this array form a tree where each child of i is at 2i and 2i+1.   Each child is 
-    // less than or equal to its parent.  
-    private DataItem[] m_heap;
-
-    #endregion
-}
-
-/// <summary>
-/// This class is responsible for taking a graph and generating a smaller graph that
-/// is a reasonable proxy.   In particular
-///     
-///     1) A spanning tree is formed, and if a node is selected so are all its 
-///        parents in that spanning tree.
-///        
-///     2) We try hard to keep scale each object type by the count by which the whole
-///        graph was reduced.  
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class GraphSampler
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public GraphSampler(MemoryGraph graph, int targetNodeCount, TextWriter log)
-    {
-        m_graph = graph;
-        m_log = log;
-        m_targetNodeCount = targetNodeCount;
-        m_filteringRatio = (float)graph.NodeCount / targetNodeCount;
-        m_nodeStorage = m_graph.AllocNodeStorage();
-        m_childNodeStorage = m_graph.AllocNodeStorage();
-        m_nodeTypeStorage = m_graph.AllocTypeNodeStorage();
-    }
-
-    /// <summary>
-    /// Creates a new graph from 'graph' which has the same type statistics as the original
-    /// graph but keeps the node count roughly at 'targetNodeCount'
-    /// </summary>
-    public MemoryGraph GetSampledGraph()
-    {
-        m_log.WriteLine("************* SAMPLING GRAPH TO REDUCE SIZE ***************");
-        m_log.WriteLine("Original graph object count {0:n0}, targetObjectCount {1:n0} targetRatio {2:f2}", m_graph.NodeCount, m_targetNodeCount, m_filteringRatio);
-        m_log.WriteLine("Original graph Size MB {0:n0} TypeCount {1:n0}", m_graph.TotalSize, m_graph.NodeTypeCount);
-
-        // Get the spanning tree
-        m_spanningTree = new SpanningTree(m_graph, m_log);
-        m_spanningTree.ForEach(null);
-
-        // Make the new graph 
-        m_newGraph = new MemoryGraph(m_targetNodeCount + m_graph.NodeTypeCount * 2);
-        m_newGraph.Is64Bit = m_graph.Is64Bit;
-
-        // Initialize the object statistics
-        m_statsByType = new SampleStats[m_graph.NodeTypeCount];
-        for (int i = 0; i < m_statsByType.Length; i++)
-        {
-            m_statsByType[i] = new SampleStats();
-        }
-
-        // And initialize the mapping from old nodes to new nodes.  (TODO: this can be a hash table to save size?  )
-        m_newIndex = new NodeIndex[m_graph.NodeCount];
-        for (int i = 0; i < m_newIndex.Length; i++)
-        {
-            m_newIndex[i] = NodeIndex.Invalid;
-        }
-
-        ValidateStats(false);
-
-        VisitNode(m_graph.RootIndex, true, false); // visit the root for sure.  
-        // Sample the nodes, trying to keep the 
-        for (NodeIndex nodeIdx = 0; nodeIdx < m_graph.NodeIndexLimit; nodeIdx++)
-        {
-            VisitNode(nodeIdx, false, false);
-        }
-
-        ValidateStats(true);
-
-        // See if we need to flesh out the potential node to become truly sampled node to hit our quota.  
-        for (NodeIndex nodeIdx = 0; nodeIdx < (NodeIndex)m_newIndex.Length; nodeIdx++)
-        {
-            var newIndex = m_newIndex[(int)nodeIdx];
-            if (newIndex == PotentialNode)
-            {
-                var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-                var stats = m_statsByType[(int)node.TypeIndex];
-                int quota = (int)(stats.TotalCount / m_filteringRatio + .5);
-                int needed = quota - stats.SampleCount;
-                if (needed > 0)
-                {
-                    // If we have not computed the frequency of sampling do it now.  
-                    if (stats.SkipFreq == 0)
-                    {
-                        var available = stats.PotentialCount - stats.SampleCount;
-                        Debug.Assert(0 <= available);
-                        Debug.Assert(needed <= available);
-                        stats.SkipFreq = Math.Max(available / needed, 1);
-                    }
-
-                    // Sample every Nth time.  
-                    stats.SkipCtr++;
-                    if (stats.SkipFreq <= stats.SkipCtr)
-                    {
-                        // Sample a new node 
-                        m_newIndex[(int)nodeIdx] = m_newGraph.CreateNode();
-                        stats.SampleCount++;
-                        stats.SampleMetric += node.Size;
-                        stats.SkipCtr = 0;
-                    }
-                }
-            }
-        }
-
-        // OK now m_newIndex tell us which nodes we want.  actually define the selected nodes. 
-
-        // Initialize the mapping from old types to new types.
-        m_newTypeIndexes = new NodeTypeIndex[m_graph.NodeTypeCount];
-        for (int i = 0; i < m_newTypeIndexes.Length; i++)
-        {
-            m_newTypeIndexes[i] = NodeTypeIndex.Invalid;
-        }
-
-        GrowableArray<NodeIndex> children = new GrowableArray<NodeIndex>(100);
-        for (NodeIndex nodeIdx = 0; nodeIdx < (NodeIndex)m_newIndex.Length; nodeIdx++)
-        {
-            // Add all sampled nodes to the new graph.  
-            var newIndex = m_newIndex[(int)nodeIdx];
-            if (IsSampledNode(newIndex))
-            {
-                var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-                // Get the children that are part of the sample (ignore ones that are filter)
-                children.Clear();
-                for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-                {
-                    var newChildIndex = m_newIndex[(int)childIndex];
-                    if (0 <= newChildIndex)                 // the child is not filtered out. 
-                    {
-                        children.Add(newChildIndex);
-                    }
-                }
-                // define the node
-                var newTypeIndex = GetNewTypeIndex(node.TypeIndex);
-                m_newGraph.SetNode(newIndex, newTypeIndex, node.Size, children);
-                m_newGraph.SetAddress(newIndex, m_graph.GetAddress(nodeIdx));
-            }
-        }
-
-        ValidateStats(true, true);
-
-        // Set the root.
-        m_newGraph.RootIndex = m_newIndex[(int)m_graph.RootIndex];
-        Debug.Assert(0 <= m_newGraph.RootIndex);            // RootIndex in the tree.  
-
-        m_newGraph.AllowReading();
-
-        // At this point we are done.  The rest is just to report the result to the user.  
-
-        // Sort the m_statsByType
-        var sortedTypes = new int[m_statsByType.Length];
-        for (int i = 0; i < sortedTypes.Length; i++)
-        {
-            sortedTypes[i] = i;
-        }
-
-        Array.Sort(sortedTypes, delegate (int x, int y)
-        {
-            var ret = m_statsByType[y].TotalMetric.CompareTo(m_statsByType[x].TotalMetric);
-            return ret;
-        });
-
-        m_log.WriteLine("Stats of the top types (out of {0:n0})", m_newGraph.NodeTypeCount);
-        m_log.WriteLine("OrigSizeMeg SampleSizeMeg   Ratio   |   OrigCnt  SampleCnt    Ratio   | Ave Size | Type Name");
-        m_log.WriteLine("---------------------------------------------------------------------------------------------");
-
-        for (int i = 0; i < Math.Min(m_statsByType.Length, 30); i++)
-        {
-            int typeIdx = sortedTypes[i];
-            NodeType type = m_graph.GetType((NodeTypeIndex)typeIdx, m_nodeTypeStorage);
-            var stats = m_statsByType[typeIdx];
-
-            m_log.WriteLine("{0,12:n6} {1,11:n6}  {2,9:f2} | {3,10:n0} {4,9:n0}  {5,9:f2} | {6,8:f0} | {7}",
-                stats.TotalMetric / 1000000.0, stats.SampleMetric / 1000000.0, stats.SampleMetric == 0 ? 0.0 : (double)stats.TotalMetric / stats.SampleMetric,
-                stats.TotalCount, stats.SampleCount, stats.SampleCount == 0 ? 0.0 : (double)stats.TotalCount / stats.SampleCount,
-                (double)stats.TotalMetric / stats.TotalCount, type.Name);
-        }
-
-        m_log.WriteLine("Sampled Graph node count {0,11:n0} (reduced by {1:f2} ratio)", m_newGraph.NodeCount,
-            (double)m_graph.NodeCount / m_newGraph.NodeCount);
-        m_log.WriteLine("Sampled Graph type count {0,11:n0} (reduced by {1:f2} ratio)", m_newGraph.NodeTypeCount,
-            (double)m_graph.NodeTypeCount / m_newGraph.NodeTypeCount);
-        m_log.WriteLine("Sampled Graph node size  {0,11:n0} (reduced by {1:f2} ratio)", m_newGraph.TotalSize,
-            (double)m_graph.TotalSize / m_newGraph.TotalSize);
-        return m_newGraph;
-    }
-
-    /// <summary>
-    /// returns an array of scaling factors.  This array is indexed by the type index of
-    /// the returned graph returned by GetSampledGraph.   If the sampled count for that type multiplied
-    /// by this scaling factor, you end up with the count for that type of the original unsampled graph.  
-    /// </summary>
-    public float[] CountScalingByType
-    {
-        get
-        {
-            var ret = new float[m_newGraph.NodeTypeCount];
-            for (int i = 0; i < m_statsByType.Length; i++)
-            {
-                var newTypeIndex = MapTypeIndex((NodeTypeIndex)i);
-                if (newTypeIndex != NodeTypeIndex.Invalid)
-                {
-                    float scale = 1;
-                    if (m_statsByType[i].SampleMetric != 0)
-                    {
-                        scale = (float)((double)m_statsByType[i].TotalMetric / m_statsByType[i].SampleMetric);
-                    }
-
-                    ret[(int)newTypeIndex] = scale;
-                }
-            }
-            for (int i = 1; i < ret.Length; i++)
-            {
-                Debug.Assert(0 < ret[i] && ret[i] <= float.MaxValue);
-            }
-
-            return ret;
-        }
-    }
-
-    /// <summary>
-    /// Maps 'oldTypeIndex' to its type index in the output graph
-    /// </summary>
-    /// <returns>New type index, will be Invalid if the type is not in the output graph</returns>
-    public NodeTypeIndex MapTypeIndex(NodeTypeIndex oldTypeIndex)
-    {
-        return m_newTypeIndexes[(int)oldTypeIndex];
-    }
-
-    /// <summary>
-    /// Maps 'oldNodeIndex' to its new node index in the output graph
-    /// </summary>
-    /// <returns>New node index, will be less than 0 if the node is not in the output graph</returns>
-    public NodeIndex MapNodeIndex(NodeIndex oldNodeIndex)
-    {
-        return m_newIndex[(int)oldNodeIndex];
-    }
-
-    #region private
-    /// <summary>
-    /// Visits 'nodeIdx', if already visited, do nothing.  If unvisited determine if 
-    /// you should add this node to the graph being built.   If 'mustAdd' is true or
-    /// if we need samples it keep the right sample/total ratio, then add the sample.  
-    /// </summary>
-    private void VisitNode(NodeIndex nodeIdx, bool mustAdd, bool doNotAddAncestors)
-    {
-        var newNodeIdx = m_newIndex[(int)nodeIdx];
-        // If this node has been selected already, there is nothing to do.    
-        if (IsSampledNode(newNodeIdx))
-        {
-            return;
-        }
-        // If we have visited this node and reject it and we are not forced to add it, we are done.
-        if (newNodeIdx == RejectedNode && !mustAdd)
-        {
-            return;
-        }
-
-        Debug.Assert(newNodeIdx == NodeIndex.Invalid || newNodeIdx == PotentialNode || (newNodeIdx == RejectedNode && mustAdd));
-
-        var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-        var stats = m_statsByType[(int)node.TypeIndex];
-
-        // If we have never seen this node before, add to our total count.  
-        if (newNodeIdx == NodeIndex.Invalid)
-        {
-            if (stats.TotalCount == 0)
-            {
-                m_numDistictTypes++;
-            }
-
-            stats.TotalCount++;
-            stats.TotalMetric += node.Size;
-        }
-
-        // Also insure that if there are a large number of types, that we sample them at least some. 
-        if (stats.SampleCount == 0 && !mustAdd && (m_numDistictTypesWithSamples + .5F) * m_filteringRatio <= m_numDistictTypes)
-        {
-            mustAdd = true;
-        }
-
-        // We sample if we are forced (it is part of a parent chain), we need it to 
-        // mimic the the original statistics, or if it is a large object (we include 
-        // all large objects, since the affect overall stats so much).  
-        if (mustAdd ||
-            (stats.PotentialCount + .5f) * m_filteringRatio <= stats.TotalCount ||
-            85000 < node.Size)
-        {
-            if (stats.SampleCount == 0)
-            {
-                m_numDistictTypesWithSamples++;
-            }
-
-            stats.SampleCount++;
-            stats.SampleMetric += node.Size;
-            if (newNodeIdx != PotentialNode)
-            {
-                stats.PotentialCount++;
-            }
-
-            m_newIndex[(int)nodeIdx] = m_newGraph.CreateNode();
-
-            // Add all direct children as potential nodes (Potential nodes I can add without adding any other node)
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                var newChildIndex = m_newIndex[(int)childIndex];
-                // Already a sampled or potential node.  Nothing to do.  
-                if (IsSampledNode(newChildIndex) || newChildIndex == PotentialNode)
-                {
-                    continue;
-                }
-
-                var childNode = m_graph.GetNode(childIndex, m_childNodeStorage);
-                var childStats = m_statsByType[(int)childNode.TypeIndex];
-
-                if (newChildIndex == NodeIndex.Invalid)
-                {
-                    if (stats.TotalCount == 0)
-                    {
-                        m_numDistictTypes++;
-                    }
-
-                    childStats.TotalCount++;
-                    childStats.TotalMetric += childNode.Size;
-                }
-                else
-                {
-                    Debug.Assert(newChildIndex == RejectedNode);
-                }
-
-                m_newIndex[(int)childIndex] = PotentialNode;
-                childStats.PotentialCount++;
-            }
-
-            // For all ancestors, require them to be in the list
-            if (!doNotAddAncestors)
-            {
-                for (; ; )
-                {
-                    nodeIdx = m_spanningTree.Parent(nodeIdx);
-                    if (nodeIdx == NodeIndex.Invalid || m_newIndex.Length == (int)nodeIdx) // The last index represents the 'orphan' node.  
-                    {
-                        break;
-                    }
-
-                    // Indicate that you should not add ancestors (since I will do this).  
-                    // We do the adding in a loop (rather than letting recursion do it) to avoid stack overflows
-                    // for long chains of objects.  
-                    VisitNode(nodeIdx, true, true);
-                }
-            }
-        }
-        else
-        {
-            if (newNodeIdx != PotentialNode)
-            {
-                m_newIndex[(int)nodeIdx] = RejectedNode;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Maps 'oldTypeIndex' to its type index in the output graph
-    /// </summary>
-    /// <param name="oldTypeIndex"></param>
-    /// <returns></returns>
-    private NodeTypeIndex GetNewTypeIndex(NodeTypeIndex oldTypeIndex)
-    {
-        var ret = m_newTypeIndexes[(int)oldTypeIndex];
-        if (ret == NodeTypeIndex.Invalid)
-        {
-            var oldType = m_graph.GetType(oldTypeIndex, m_nodeTypeStorage);
-            ret = m_newGraph.CreateType(oldType.Name, oldType.ModuleName, oldType.Size);
-            m_newTypeIndexes[(int)oldTypeIndex] = ret;
-        }
-        return ret;
-    }
-
-
-    [Conditional("DEBUG")]
-    private void ValidateStats(bool allNodesVisited, bool completed = false)
-    {
-        var statsCheckByType = new SampleStats[m_statsByType.Length];
-        for (int i = 0; i < statsCheckByType.Length; i++)
-        {
-            statsCheckByType[i] = new SampleStats();
-        }
-
-        int total = 0;
-        long totalSize = 0;
-        int sampleTotal = 0;
-        var typeStorage = m_graph.AllocTypeNodeStorage();
-        for (NodeIndex nodeIdx = 0; nodeIdx < (NodeIndex)m_newIndex.Length; nodeIdx++)
-        {
-            var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-            var stats = statsCheckByType[(int)node.TypeIndex];
-            var type = node.GetType(typeStorage);
-            var typeName = type.Name;
-            var newNodeIdx = m_newIndex[(int)nodeIdx];
-
-            if (newNodeIdx == NodeIndex.Invalid)
-            {
-                // We should have visited every node, so there should be no Invalid nodes. 
-                Debug.Assert(!allNodesVisited);
-            }
-            else
-            {
-                total++;
-                stats.TotalCount++;
-                stats.TotalMetric += node.Size;
-                totalSize += node.Size;
-                Debug.Assert(node.Size != 0 || typeName.StartsWith('[') || typeName == "UNDEFINED");
-                if (IsSampledNode(newNodeIdx) || newNodeIdx == PotentialNode)
-                {
-                    if (nodeIdx != m_graph.RootIndex)
-                    {
-                        Debug.Assert(IsSampledNode(m_spanningTree.Parent(nodeIdx)));
-                    }
-
-                    stats.PotentialCount++;
-                    if (IsSampledNode(newNodeIdx))
-                    {
-                        stats.SampleCount++;
-                        sampleTotal++;
-                        stats.SampleMetric += node.Size;
-                    }
-                }
-                else
-                {
-                    Debug.Assert(newNodeIdx == RejectedNode);
-                }
-            }
-            statsCheckByType[(int)node.TypeIndex] = stats;
-        }
-
-        float[] scalings = null;
-        if (completed)
-        {
-            scalings = CountScalingByType;
-        }
-
-        for (NodeTypeIndex typeIdx = 0; typeIdx < m_graph.NodeTypeIndexLimit; typeIdx++)
-        {
-            _ = m_graph.GetType(typeIdx, typeStorage);
-            var statsCheck = statsCheckByType[(int)typeIdx];
-            var stats = m_statsByType[(int)typeIdx];
-
-            Debug.Assert(stats.TotalMetric == statsCheck.TotalMetric);
-            Debug.Assert(stats.TotalCount == statsCheck.TotalCount);
-            Debug.Assert(stats.SampleCount == statsCheck.SampleCount);
-            Debug.Assert(stats.SampleMetric == statsCheck.SampleMetric);
-            Debug.Assert(stats.PotentialCount == statsCheck.PotentialCount);
-
-            Debug.Assert(stats.PotentialCount <= statsCheck.TotalCount);
-            Debug.Assert(stats.SampleCount <= statsCheck.PotentialCount);
-
-            // We should be have at least m_filterRatio of Potential objects 
-            Debug.Assert(!((stats.PotentialCount + .5f) * m_filteringRatio <= stats.TotalCount));
-
-            // If we completed, then we converted potentials to true samples.   
-            if (completed)
-            {
-                Debug.Assert(!((stats.SampleCount + .5f) * m_filteringRatio <= stats.TotalCount));
-
-                // Make sure that scalings that we finally output were created correctly
-                if (stats.SampleMetric > 0)
-                {
-                    var newTypeIdx = MapTypeIndex(typeIdx);
-                    var estimatedTotalMetric = scalings[(int)newTypeIdx] * stats.SampleMetric;
-                    Debug.Assert(Math.Abs(estimatedTotalMetric - stats.TotalMetric) / stats.TotalMetric < .01);
-                }
-            }
-
-            if (stats.SampleCount == 0)
-            {
-                Debug.Assert(stats.SampleMetric == 0);
-            }
-
-            if (stats.TotalMetric == 0)
-            {
-                Debug.Assert(stats.TotalMetric == 0);
-            }
-        }
-
-        if (allNodesVisited)
-        {
-            Debug.Assert(total == m_graph.NodeCount);
-            // TODO The assert should be Debug.Assert(totalSize == m_graph.TotalSize);
-            // but we have to give a 1% error margin to get things passing. Fix this.
-            Debug.Assert(Math.Abs(totalSize - m_graph.TotalSize) / totalSize < .01);
-        }
-        Debug.Assert(sampleTotal == m_newGraph.NodeCount);
-    }
-
-    private sealed class SampleStats
-    {
-        public int TotalCount;          // The number of objects of this type in the original graph
-        public int SampleCount;         // The number of objects of this type we have currently added to the new graph
-        public int PotentialCount;      // SampleCount + The number of objects of this type that can be added without needing to add other nodes
-        public long TotalMetric;
-        public long SampleMetric;
-        public int SkipFreq;          // When sampling potentials, take every Nth one where this is the N
-        public int SkipCtr;             // This remembers our last N.  
-    };
-
-    /// <summary>
-    /// This value goes in the m_newIndex[].   If we accept the node into the sampled graph, we put the node
-    /// index in the NET graph in m_newIndex.   If we reject the node we use the special RejectedNode value
-    /// below
-    /// </summary>
-    private const NodeIndex RejectedNode = (NodeIndex)(-2);
-
-    /// <summary>
-    /// This value also goes in m_newIndex[].   If we can add this node without needing to add any other nodes
-    /// to the new graph (that is it is one hop from an existing accepted node, then we mark it specially as
-    /// a PotentialNode).   We add these in a second pass over the data.  
-    /// </summary>
-    private const NodeIndex PotentialNode = (NodeIndex)(-3);
-
-    private bool IsSampledNode(NodeIndex nodeIdx) { return 0 <= nodeIdx; }
-
-    private MemoryGraph m_graph;
-    private int m_targetNodeCount;
-    private TextWriter m_log;
-    private Node m_nodeStorage;
-    private Node m_childNodeStorage;
-    private NodeType m_nodeTypeStorage;
-    private float m_filteringRatio;
-    private SampleStats[] m_statsByType;
-    private int m_numDistictTypesWithSamples;
-    private int m_numDistictTypes;
-    private NodeIndex[] m_newIndex;
-    private NodeTypeIndex[] m_newTypeIndexes;
-    private SpanningTree m_spanningTree;
-    private MemoryGraph m_newGraph;
-    #endregion
-}
-
 #if false
 namespace Experimental
 {
     /// <summary>
     /// code:PagedGrowableArray is an array (has an index operation) but can efficiently represent
-    /// either very large arrays as well as sparse arrays.  
+    /// either very large arrays as well as sparse arrays.
     /// </summary>
     public struct PagedGrowableArray<T>
     {
@@ -2710,7 +1390,7 @@ namespace Experimental
             }
         }
         /// <summary>
-        /// Append the value to the end of the array.  
+        /// Append the value to the end of the array.
         /// </summary>
         /// <param name="value"></param>
         public void Add(T value)
@@ -2755,7 +1435,7 @@ namespace Experimental
             }
         }
         /// <summary>
-        /// Append the value to the end of the array.  
+        /// Append the value to the end of the array.
         /// </summary>
         /// <param name="value"></param>
         public void Add(long value)
@@ -2786,10 +1466,10 @@ namespace Experimental
         }
 
         /// <summary>
-        /// A page represents 256 entries in the table.   For each page we remember a 'm_baseValue' and 
+        /// A page represents 256 entries in the table.   For each page we remember a 'm_baseValue' and
         /// we delta encode.  If the offset fit 15 bits you simply add the offset to the base value
         /// Otherwise what is in the table is an offset into the 'm_compressedValues' blob and the offset
-        /// is encoded as a variable length signed number.  
+        /// is encoded as a variable length signed number.
         /// </summary>
         class Page : IFastSerializable
         {
@@ -2864,10 +1544,10 @@ namespace Experimental
                 }
             }
 
-            long m_baseValue;                  // All values are relative to this.  
+            long m_baseValue;                  // All values are relative to this.
             short[] m_indexOrOffset;           // table of value (either offsets or indexes into the compressed blobs)
 
-            byte[] m_compressedValues;          // If all the values are not within 32K of the base, then store them here.  
+            byte[] m_compressedValues;          // If all the values are not within 32K of the base, then store them here.
             int m_compressedValuesIndex;        // Next place to write to in m_compressedValues
 #endregion
         }
